--- src/main/java/OrderReposoinity.java
@@ -1,5 +1,0 @@
-
-import org.springframework.data.jpa.repository.JpaRepository;
-public interface OrderReposoinity  extends JpaRepository<Order,Long> {
-
-}
