+++ src/main/java/com/qa/OrderReposoinity.java
@@ -1,0 +1,5 @@
+package com.qa;
+import org.springframework.data.jpa.repository.JpaRepository;
+public interface OrderReposoinity  extends JpaRepository<OrderFilm,Long> {
+
+}
