+++ src/main/java/com/qa/OrderTicket.java
@@ -1,0 +1,42 @@
+//package com.qa;
+//
+//
+//
+//import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+//
+//import javax.persistence.*;
+//
+//@Entity
+//@JsonIgnoreProperties({"hibernateLazyIntializer","handler"})
+//public class OrderTicket {
+//     @Id
+//     private long ticketId;
+//     //@ManyToOne
+//     //@JoinColumn(name="orderIdf")
+//     private OrderFilm order;
+//
+//    public OrderTicket(long ticketId, OrderFilm order) {
+//        this.ticketId = ticketId;
+//        this.order = order;
+//    }
+//
+//    public long getTicketId() {
+//        return this.ticketId;
+//    }
+//
+//    public void setTicketId(long ticketId) {
+//        this.ticketId = ticketId;
+//    }
+//
+//    public OrderFilm getOrder() {
+//        return this.order;
+//    }
+//
+//    public void setOrder(OrderFilm order) {
+//        this.order = order;
+//    }
+//
+//
+//}
+
+
