+++ src/main/resources/application.properties
@@ -16,3 +16,4 @@ spring.datasource.username=sa
 spring.datasource.password=sa
 
 spring.datasource.driver-class-name=org.h2.Driver
+#spring.jpa.hibernate.ddl-auto=none
